Complete Item 9
This item was okay. I was able to make my functions simple and readable, but I was amazed as to how little lines of code you needed for this problem. The solution only had 2 lines of code. Not only that, it was easy to understand as well.

This shows how much I need to be familiar with JS functions.