Complete Item 12
I had trouble with this. I knew that the strightforward approach would solve it easily, so I wanted to use the reduce function. I added an age attribute for tracking with ease, but the problem was building the reduce function.

It was simple, yet I overlooked it. I learned that if there are braces after the lambda function, I need to put the return keyword. Otherwise, I can omit it. No wonder the reduce function kept on returning undefined. Also, the reduce function can handle objects.